write-up: desurvey
started a blog! Something like a blog! Well, I wrote some stuff on a subject I like and couldn't find a nice reference, so there it is. The sole reference (for now) includes a link to the internet archive, in case the leapfrog3d blog cease to exist. Exciting times!